V0.8
lesson 8, refact code, MVC, delete 6,7 lessons homework
