+++ RevinaUnityHomeWork/Assets/Scripts/Controllers/BadBonus.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 2311b63f04938ac40bf87068314dc8bf
+guid: 4959a44c71c254b4b89beda40bc86d44
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
