+++ RevinaUnityHomeWork/Assets/Scripts/Controllers/GoodBonus.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: fa97fa5a05041be418913fbd34f28812
+guid: 89bf5e6a4422ab84f8cd9904dd76f514
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
