+++ RevinaUnityHomeWork/Assets/Scripts/Controllers/GoodBonusComparer.cs
@@ -1,6 +1,4 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
+﻿using System.Collections.Generic;
 
 namespace Geekbrains
 {
