+++ RevinaUnityHomeWork/Assets/Scripts/Controllers/GoodBonusComparer.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 183386dfd1ec12945b9b76109457deaf
+guid: b59049f3d5235ff4eac3ae2b725d6cc3
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
