+++ RevinaUnityHomeWork/Assets/Scripts/Controllers/ListInteractableObject.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: c2ee5d46a9580074bb639dba1fa4ff2f
+guid: e0000c8633232d84f9c1cf49c14fe2f8
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
