--- RevinaUnityHomeWork/Assets/Scripts/HomeWorkLinq/StringExtension.cs
@@ -1,22 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace ExtentionMethodTest
-{
-    public static class StringExtension
-    {
-        public static int CharCount(this string str, char c)
-        {
-            int counter = 0;
-            for (int i = 0; i < str.Length; i++)
-            {
-                if (str[i] == c)
-                    counter++;
-            }
-            return counter;
-        }
-    }
-}
