+++ RevinaUnityHomeWork/Assets/Scripts/Model/BadBonusModel.cs
@@ -1,0 +1,15 @@
+﻿using System;
+
+namespace Geekbrains
+{
+    public class BadBonusModel : BadBonus
+    {
+        private event EventHandler<CaughtPlayerEventArgs> _caughtPlayer;
+
+        public event EventHandler<CaughtPlayerEventArgs> CaughtPlayer
+        {
+            add { _caughtPlayer += value; }
+            remove { _caughtPlayer -= value; }
+        }
+    }
+}
