+++ RevinaUnityHomeWork/Assets/Scripts/Model/BadBonusModel.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 6cd77bbcd5c094d458b7e7f4b3e3d9dc
+guid: 821a10bbe3135e64483f64533d95c4d2
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
