+++ RevinaUnityHomeWork/Assets/Scripts/Model/CaughtPlayerEventArgs.cs
