+++ RevinaUnityHomeWork/Assets/Scripts/Model/CaughtPlayerEventArgs.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3199d91335b86c24fa7ca374cb712bfb
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
