+++ RevinaUnityHomeWork/Assets/Scripts/Model/InteractObjectsView.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 7a437504a2cf31041a9cddb191f8eb92
+guid: d1f7f9ffc2710204c9ef8892a499c617
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
