+++ RevinaUnityHomeWork/Assets/Scripts/Model/InteractiveObjects.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 03ded6613b5fb1d43ad696778110c9d6
+guid: ed192dd79e0fc6445b8f8091ec891114
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
