+++ RevinaUnityHomeWork/Assets/Scripts/Model/PlayerBall.cs
@@ -1,6 +1,4 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
+﻿using UnityEngine;
 
 namespace Geekbrains
 {
