+++ RevinaUnityHomeWork/Assets/Scripts/View/BadBonusView.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 3f037c849aefe664e84fb5a526461b7f
+guid: ea6c5145b031cc24baf309501916742a
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
