+++ RevinaUnityHomeWork/Assets/Scripts/View/GoodBonusView.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 58683d31a67427546b1bce98cdb409cd
+guid: f3a05b82b3197d047b832a7263723fe0
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
