+++ RevinaUnityHomeWork/Assets/Scripts/View/MapView.cs
@@ -1,5 +1,4 @@
-﻿using System.Collections;
-using UnityEngine;
+﻿using UnityEngine;
 
 namespace Geekbrains
 {
